Bot Delay and Failed Job Handling
Bot busy timeouts and delays are handled within the bot class on a job
fail. For delays, it sends an extra property 'delay' to the resolve,
which in turn delays the call of the done function. This ensures that
the requestor of the data gets it right away without having to wait for
the bot busy state to reset due to request delay.

Since we can't pass non-stringifiable objects into the Kue job data,
adding a bot obj reference to the job data a bot is handling is not
possible. Creating a separate dictionary with them is possible though
and would alleviate 2 redundant timeouts in the bot obj itself.