Removes Kue By Implementing Custom Queue
By removing the need for Kue, we don't have to handle done objects the
same way and do TTL handling within the bot and the promise.

Completely removes the need to handle done objects in `res_controller`,
Redis, and fixes small bugs in relation to the implementation of Kue.

We didn't need persistance across restarts by using Kue anyways, this is
a simple "drop-in" replacement.